Revert "Decision tree"
