--- src/components/MontyTreeTabs/MontyTreeTabs.css
@@ -1,62 +1,0 @@
-.monty-tree-tabs {
-  width: 100%;
-}
-
-.tab-controls {
-  display: flex;
-  justify-content: center;
-  gap: 4px;
-  margin: 20px 0 30px 0;
-  padding: 0;
-  background-color: transparent;
-  flex-wrap: wrap;
-}
-
-.tab-button {
-  padding: 16px 24px;
-  background-color: #f8f9fa;
-  color: #6c757d;
-  border: 2px solid #e9ecef;
-  border-radius: 8px 8px 0 0;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  font-family: inherit;
-  text-align: center;
-  min-width: 120px;
-  position: relative;
-}
-
-.tab-button:hover {
-  background-color: #e9ecef;
-  color: #495057;
-  transform: translateY(-2px);
-}
-
-.tab-button.active {
-  background-color: #007bff;
-  color: white;
-  border-color: #007bff;
-  transform: translateY(-4px);
-  box-shadow: 0 4px 12px rgba(0, 123, 255, 0.3);
-}
-
-.tab-button.active:hover {
-  background-color: #0056b3;
-  border-color: #0056b3;
-}
-
-.tab-label {
-  font-weight: 600;
-  font-size: 1.1em;
-  margin-bottom: 4px;
-}
-
-.tab-description {
-  font-size: 0.85em;
-  opacity: 0.8;
-  line-height: 1.2;
-}
-
-.tab-button.active .tab-description {
-  opacity: 0.9;
-}
