+++ src/pages/EducationPage/EducationPage.css
@@ -1,3 +1,4 @@
+
 .education-container {
   padding: 2rem;
   font-family: sans-serif;
@@ -23,4 +24,239 @@
 
 #resourcesTitle{
   text-align: center;
+}
+
+.tree-visualization-section {
+  margin-top: 40px;
+  padding: 30px 20px;
+  border-radius: 8px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  clear: both;
+  display: block;
+  float: none;
+  position: relative;
+}
+
+.tree-visualization-section h1 {
+  text-align: center;
+  margin-bottom: 10px;
+  font-size: 2em;
+  font-weight: 600;
+  margin-top: 0;
+}
+
+.tree-visualization-section p {
+  text-align: center;
+  font-size: 1em;
+  margin-bottom: 25px;
+  line-height: 1.6;
+}
+
+.tree-controls {
+  display: flex;
+  justify-content: center;
+  gap: 15px;
+  margin: 25px 0;
+  flex-wrap: wrap;
+}
+
+.control-btn {
+  padding: 10px 20px;
+  background-color: #007bff;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  font-weight: 500;
+  font-size: 14px;
+  transition: background-color 0.3s ease;
+  text-transform: none;
+  letter-spacing: normal;
+}
+
+.control-btn:hover {
+  background-color: #0056b3;
+  transform: none;
+  box-shadow: none;
+}
+
+.control-btn:active {
+  transform: translateY(1px);
+}
+
+.tree-visualization {
+  display: flex;
+  justify-content: center;
+  margin: 30px 0;
+  border-radius: 6px;
+  padding: 15px;
+  overflow: visible;
+  height: 25em;
+}
+
+.tree-visualization svg {
+  border-radius: 4px;
+  max-width: 100%;
+  height: auto;          
+  min-height: 320px;     
+}
+
+.tree-connection {
+  stroke: #666;
+  stroke-width: 2;
+  opacity: 0.7;
+}
+
+.tree-node-circle {
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.neutral-node {
+  fill: #f8f9fa;
+  stroke: #6c757d;
+  stroke-width: 2;
+}
+
+.neutral-node:hover {
+  fill: #e9ecef;
+  stroke: #495057;
+  stroke-width: 2;
+}
+
+.win-node {
+  fill: #d4edda;
+  stroke: #28a745;
+  stroke-width: 2;
+}
+
+.win-node:hover {
+  fill: #c3e6cb;
+  stroke: #1e7e34;
+  stroke-width: 2;
+}
+
+.lose-node {
+  fill: #f8d7da;
+  stroke: #dc3545;
+  stroke-width: 2;
+}
+
+.lose-node:hover {
+  fill: #f5c6cb;
+  stroke: #bd2130;
+  stroke-width: 2;
+}
+
+.expanded {
+  fill: #cce7ff;
+  stroke: #007bff;
+  stroke-width: 2;
+}
+
+.expanded:hover {
+  fill: #b3d9ff;
+  stroke: #0056b3;
+  stroke-width: 2;
+}
+
+.node-label {
+  fill: white;
+  font-family: Arial, sans-serif;
+  font-size: 11px;
+  font-weight: 500;
+  pointer-events: none;
+}
+
+.node-probability {
+  fill: white;
+  font-family: Arial, sans-serif;
+  font-size: 10px;
+  font-style: italic;
+  pointer-events: none;
+}
+
+.expand-indicator {
+  fill: #333;
+  font-family: Arial, sans-serif;
+  font-size: 12px;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.outcome-text {
+  font-family: Arial, sans-serif;
+  font-size: 9px;
+  font-weight: bold;
+  pointer-events: none;
+  text-transform: uppercase;
+}
+
+.win-text {
+  fill: #28a745;
+}
+
+.lose-text {
+  fill: #dc3545;
+}
+
+.win-rates {
+  margin: 30px 0 0 0;
+  padding: 20px;
+  border-radius: 6px;
+}
+
+.win-rates h4 {
+  margin-top: 0;
+  margin-bottom: 20px;
+  text-align: center;
+  font-size: 1.3em;
+  font-weight: 600;
+}
+
+.rate-comparison {
+  display: flex;
+  gap: 20px;
+  margin-top: 15px;
+  flex-wrap: wrap;
+}
+
+.rate-item {
+  padding: 16px 20px;
+  border-radius: 6px;
+  font-weight: 600;
+  flex: 1;
+  min-width: 250px;
+  text-align: center;
+  border: 1px solid;
+}
+
+
+@media (max-width: 768px) {
+  .tree-visualization-section {
+    margin-top: 30px;
+    padding: 20px 15px;
+  }
+  
+  .tree-visualization-section h1 {
+    font-size: 1.6em;
+  }
+  
+  .tree-visualization svg {
+    min-width: 600px;
+    height: 350px;
+  }
+  
+  .rate-comparison {
+    flex-direction: column;
+  }
+  
+  .rate-item {
+    margin-bottom: 10px;
+  }
+  
+  .control-btn {
+    padding: 8px 16px;
+    font-size: 13px;
+  }
 }
